Add ability to register name after setting up bus
When using dbus activation, as soon as the bus name is registered then
any pending messages are sent. If the name is registered before the
objects are registered on the bus then the messages can be ignored. By
separating the bus creation from the name registration then the bus can
be made fully operational before the registering of the name triggers
incoming messages.